+++ crates/mako/src/compiler.rs
@@ -237,7 +237,7 @@ impl Compiler {
             Arc::new(plugins::async_runtime::AsyncRuntimePlugin {}),
             Arc::new(plugins::emotion::EmotionPlugin {}),
             Arc::new(plugins::tree_shaking::FarmTreeShake {}),
-            Arc::new(plugins::loop_circular_dependence::LoopDetector {}),
+            Arc::new(plugins::detect_circular_dependence::LoopDetector {}),
         ];
         plugins.extend(builtin_plugins);
