+++ crates/mako/src/plugins/detect_circular_dependence.rs
