+++ crates/mako/src/plugins/mod.rs
@@ -2,13 +2,13 @@ pub mod async_runtime;
 pub mod bundless_compiler;
 pub mod context_module;
 pub mod copy;
+pub mod detect_circular_dependence;
 pub mod emotion;
 pub mod graphviz;
 pub mod hmr_runtime;
 pub mod ignore;
 pub mod import;
 pub mod invalid_webpack_syntax;
-pub mod loop_circular_dependence;
 pub mod manifest;
 pub mod minifish;
 pub mod require_context;
